Provide an alternative interface
This is a hard decision -- I need to provide an alternative interface. This
is the only compromise that I can think of.

The original problem is explained in
http://github.com/fayland/perl-lingua-han/commit/799af3edb9a5207123b70a0662f99747f094ec47

However, the fix also introduced the encoding guessing problem as explained in
http://github.com/fayland/perl-lingua-han/commit/e2736bd82ca3ac25aa694e73b8c48c4efb4805d8

To the best of my knowledge, until the Perl Encode::Guess->guess() can be
versatile enough to be bullet-proof, the only solution is not to provide a
panacea interface, but individual ones that take the guessing game out.